Add files via upload
The Charusat hostel management System is designed to show the all hostels details and its different types rooms and mess details. It included all the hostels which is near by the charusat university. A student can make a request for any type of room of any hostel by this application. This application also shows the data of particular student.